Disciplines HomeWork

+++ Blazor/BlazorAppPD_212/Models/Discipline.cs
@@ -5,10 +5,10 @@ namespace BlazorAppPD_212.Models
     public class Discipline
     {
         [Key]
-        public int discipline_id { get; set; }
+        public short discipline_id { get; set; }
         [Required]
         public string discipline_name { get; set; }
         [Required]
-        public int number_of_lessons { get; set; }
+        public byte number_of_lessons { get; set; }
     }
 }
